remove bridge
